+++ example/lib/count_down_timer_page.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:stop_watch_timer/stop_watch_timer.dart';
+import 'package:stop_watch_timer_example/rounded_button.dart';
 
 class CountDownTimerPage extends StatefulWidget {
   static Future<void> navigatorPush(BuildContext context) async {
@@ -23,8 +24,8 @@ class _State extends State<CountDownTimerPage> {
     onChange: (value) => print('onChange $value'),
     onChangeRawSecond: (value) => print('onChangeRawSecond $value'),
     onChangeRawMinute: (value) => print('onChangeRawMinute $value'),
-    onStop: () {
-      print('onStop');
+    onStopped: () {
+      print('onStopped');
     },
     onEnded: () {
       print('onEnded');
@@ -41,7 +42,8 @@ class _State extends State<CountDownTimerPage> {
     _stopWatchTimer.minuteTime.listen((value) => print('minuteTime $value'));
     _stopWatchTimer.secondTime.listen((value) => print('secondTime $value'));
     _stopWatchTimer.records.listen((value) => print('records $value'));
-    _stopWatchTimer.fetchStop.listen((value) => print('stop from stream'));
+    _stopWatchTimer.fetchStopped
+        .listen((value) => print('stopped from stream'));
     _stopWatchTimer.fetchEnded.listen((value) => print('ended from stream'));
 
     /// Can be set preset time. This case is "00:01.23".
@@ -251,16 +253,9 @@ class _State extends State<CountDownTimerPage> {
                     children: <Widget>[
                       Padding(
                         padding: const EdgeInsets.symmetric(horizontal: 4),
-                        child: ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            primary: Colors.lightBlue,
-                            onPrimary: Colors.white,
-                            shape: const StadiumBorder(),
-                          ),
-                          onPressed: () async {
-                            _stopWatchTimer.onExecute
-                                .add(StopWatchExecute.start);
-                          },
+                        child: RoundedButton(
+                          color: Colors.lightBlue,
+                          onTap: _stopWatchTimer.onStartTimer,
                           child: const Text(
                             'Start',
                             style: TextStyle(color: Colors.white),
@@ -269,16 +264,9 @@ class _State extends State<CountDownTimerPage> {
                       ),
                       Padding(
                         padding: const EdgeInsets.symmetric(horizontal: 4),
-                        child: ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            primary: Colors.green,
-                            onPrimary: Colors.white,
-                            shape: const StadiumBorder(),
-                          ),
-                          onPressed: () async {
-                            _stopWatchTimer.onExecute
-                                .add(StopWatchExecute.stop);
-                          },
+                        child: RoundedButton(
+                          color: Colors.green,
+                          onTap: _stopWatchTimer.onStopTimer,
                           child: const Text(
                             'Stop',
                             style: TextStyle(color: Colors.white),
@@ -287,16 +275,9 @@ class _State extends State<CountDownTimerPage> {
                       ),
                       Padding(
                         padding: const EdgeInsets.symmetric(horizontal: 4),
-                        child: ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            primary: Colors.red,
-                            onPrimary: Colors.white,
-                            shape: const StadiumBorder(),
-                          ),
-                          onPressed: () async {
-                            _stopWatchTimer.onExecute
-                                .add(StopWatchExecute.reset);
-                          },
+                        child: RoundedButton(
+                          color: Colors.red,
+                          onTap: _stopWatchTimer.onResetTimer,
                           child: const Text(
                             'Reset',
                             style: TextStyle(color: Colors.white),
@@ -313,15 +294,9 @@ class _State extends State<CountDownTimerPage> {
                     children: <Widget>[
                       Padding(
                         padding: const EdgeInsets.all(0).copyWith(right: 8),
-                        child: ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            primary: Colors.deepPurpleAccent,
-                            onPrimary: Colors.white,
-                            shape: const StadiumBorder(),
-                          ),
-                          onPressed: () async {
-                            _stopWatchTimer.onExecute.add(StopWatchExecute.lap);
-                          },
+                        child: RoundedButton(
+                          color: Colors.deepPurpleAccent,
+                          onTap: _stopWatchTimer.onAddLap,
                           child: const Text(
                             'Lap',
                             style: TextStyle(color: Colors.white),
@@ -338,13 +313,9 @@ class _State extends State<CountDownTimerPage> {
                     children: <Widget>[
                       Padding(
                         padding: const EdgeInsets.symmetric(horizontal: 4),
-                        child: ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            primary: Colors.pinkAccent,
-                            onPrimary: Colors.white,
-                            shape: const StadiumBorder(),
-                          ),
-                          onPressed: () async {
+                        child: RoundedButton(
+                          color: Colors.pinkAccent,
+                          onTap: () {
                             _stopWatchTimer.setPresetHoursTime(1);
                           },
                           child: const Text(
@@ -355,13 +326,9 @@ class _State extends State<CountDownTimerPage> {
                       ),
                       Padding(
                         padding: const EdgeInsets.symmetric(horizontal: 4),
-                        child: ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            primary: Colors.pinkAccent,
-                            onPrimary: Colors.white,
-                            shape: const StadiumBorder(),
-                          ),
-                          onPressed: () async {
+                        child: RoundedButton(
+                          color: Colors.pinkAccent,
+                          onTap: () {
                             _stopWatchTimer.setPresetMinuteTime(59);
                           },
                           child: const Text(
@@ -372,13 +339,9 @@ class _State extends State<CountDownTimerPage> {
                       ),
                       Padding(
                         padding: const EdgeInsets.symmetric(horizontal: 4),
-                        child: ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                            primary: Colors.pinkAccent,
-                            onPrimary: Colors.white,
-                            shape: const StadiumBorder(),
-                          ),
-                          onPressed: () async {
+                        child: RoundedButton(
+                          color: Colors.pinkAccent,
+                          onTap: () {
                             _stopWatchTimer.setPresetSecondTime(10);
                           },
                           child: const Text(
@@ -392,13 +355,9 @@ class _State extends State<CountDownTimerPage> {
                 ),
                 Padding(
                   padding: const EdgeInsets.symmetric(horizontal: 4),
-                  child: ElevatedButton(
-                    style: ElevatedButton.styleFrom(
-                      primary: Colors.pinkAccent,
-                      onPrimary: Colors.white,
-                      shape: const StadiumBorder(),
-                    ),
-                    onPressed: () async {
+                  child: RoundedButton(
+                    color: Colors.pinkAccent,
+                    onTap: () {
                       _stopWatchTimer.setPresetTime(mSec: 3599 * 1000);
                     },
                     child: const Text(
@@ -409,13 +368,9 @@ class _State extends State<CountDownTimerPage> {
                 ),
                 Padding(
                   padding: const EdgeInsets.symmetric(horizontal: 4),
-                  child: ElevatedButton(
-                    style: ElevatedButton.styleFrom(
-                      primary: Colors.pinkAccent,
-                      onPrimary: Colors.white,
-                      shape: const StadiumBorder(),
-                    ),
-                    onPressed: () async {
+                  child: RoundedButton(
+                    color: Colors.pinkAccent,
+                    onTap: () {
                       _stopWatchTimer.clearPresetTime();
                     },
                     child: const Text(